requirements.txt
python-dateutil==2.8.1
phonenumbers==8.12.31
pyyaml==5.4.1
python-dotenv==0.19.1
selenium==4.4.0
pyjwt==2.4.0
